+++ rodbus/src/server/request.rs
@@ -3,7 +3,7 @@ use crate::service::function::FunctionCode;
 use crate::service::traits::ParseRequest;
 use crate::types::{
     AddressRange, BitIterator, Indexed, ReadBitsRange, ReadRegistersRange, RegisterIterator,
-    WriteCoils, WriteMultiple, WriteRegisters,
+    WriteCoils, WriteRegisters,
 };
 use crate::util::cursor::ReadCursor;
 
@@ -72,6 +72,137 @@ impl<'a> Request<'a> {
                     RegisterIterator::parse_all(range, cursor)?,
                 )))
             }
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+
+    #[cfg(test)]
+    mod coils {
+        use crate::util::cursor::ReadCursor;
+
+        use super::super::*;
+        use crate::error::details::ADUParseError;
+        use crate::types::Indexed;
+
+        #[test]
+        fn fails_when_too_few_bytes_for_coil_byte_count() {
+            let mut cursor = ReadCursor::new(&[0x00, 0x01, 0x00, 0x08, 0x00]);
+            let err = Request::parse(FunctionCode::WriteMultipleCoils, &mut cursor)
+                .err()
+                .unwrap();
+            assert_eq!(err, ADUParseError::InsufficientBytes.into());
+        }
+
+        #[test]
+        fn fails_when_too_many_bytes_for_coil_byte_count() {
+            let mut cursor = ReadCursor::new(&[0x00, 0x01, 0x00, 0x08, 0x02]);
+            let err = Request::parse(FunctionCode::WriteMultipleCoils, &mut cursor)
+                .err()
+                .unwrap();
+            assert_eq!(err, ADUParseError::InsufficientBytes.into());
+        }
+
+        #[test]
+        fn fails_when_specified_byte_count_not_present() {
+            let mut cursor = ReadCursor::new(&[0x00, 0x01, 0x00, 0x08, 0x01]);
+            let err = Request::parse(FunctionCode::WriteMultipleCoils, &mut cursor)
+                .err()
+                .unwrap();
+            assert_eq!(err, ADUParseError::InsufficientBytes.into());
+        }
+
+        #[test]
+        fn fails_when_too_many_bytes_present() {
+            let mut cursor = ReadCursor::new(&[0x00, 0x01, 0x00, 0x08, 0x01, 0xFF, 0xFF]);
+            let err = Request::parse(FunctionCode::WriteMultipleCoils, &mut cursor)
+                .err()
+                .unwrap();
+            assert_eq!(err, ADUParseError::TrailingBytes(1).into());
+        }
+
+        #[test]
+        fn can_parse_coils() {
+            let mut cursor = ReadCursor::new(&[0x00, 0x01, 0x00, 0x03, 0x01, 0x05]);
+            let coils = match Request::parse(FunctionCode::WriteMultipleCoils, &mut cursor).unwrap()
+            {
+                Request::WriteMultipleCoils(write) => write,
+                _ => panic!("bad match"),
+            };
+
+            assert_eq!(coils.range, AddressRange::try_from(1, 3).unwrap());
+            assert_eq!(
+                coils.iterator.collect::<Vec<Indexed<bool>>>(),
+                vec![
+                    Indexed::new(1, true),
+                    Indexed::new(2, false,),
+                    Indexed::new(3, true)
+                ]
+            )
+        }
+    }
+
+    #[cfg(test)]
+    mod registers {
+        use crate::util::cursor::ReadCursor;
+
+        use super::super::*;
+        use crate::error::details::ADUParseError;
+        use crate::types::Indexed;
+
+        #[test]
+        fn fails_when_too_few_bytes_for_coil_byte_count() {
+            let mut cursor = ReadCursor::new(&[0x00, 0x01, 0x00, 0x01, 0x00]);
+            let err = Request::parse(FunctionCode::WriteMultipleRegisters, &mut cursor)
+                .err()
+                .unwrap();
+            assert_eq!(err, ADUParseError::InsufficientBytes.into());
+        }
+
+        #[test]
+        fn fails_when_too_many_bytes_for_coil_byte_count() {
+            let mut cursor = ReadCursor::new(&[0x00, 0x01, 0x00, 0x01, 0x03]);
+            let err = Request::parse(FunctionCode::WriteMultipleRegisters, &mut cursor)
+                .err()
+                .unwrap();
+            assert_eq!(err, ADUParseError::InsufficientBytes.into());
+        }
+
+        #[test]
+        fn fails_when_specified_byte_count_not_present() {
+            let mut cursor = ReadCursor::new(&[0x00, 0x01, 0x00, 0x01, 0x02, 0xFF]);
+            let err = Request::parse(FunctionCode::WriteMultipleRegisters, &mut cursor)
+                .err()
+                .unwrap();
+            assert_eq!(err, ADUParseError::InsufficientBytes.into());
+        }
+
+        #[test]
+        fn fails_when_too_many_bytes_present() {
+            let mut cursor = ReadCursor::new(&[0x00, 0x01, 0x00, 0x01, 0x02, 0xFF, 0xFF, 0xFF]);
+            let err = Request::parse(FunctionCode::WriteMultipleRegisters, &mut cursor)
+                .err()
+                .unwrap();
+            assert_eq!(err, ADUParseError::TrailingBytes(1).into());
+        }
+
+        #[test]
+        fn can_parse_registers() {
+            let mut cursor =
+                ReadCursor::new(&[0x00, 0x01, 0x00, 0x02, 0x04, 0xCA, 0xFE, 0xBB, 0xDD]);
+            let registers =
+                match Request::parse(FunctionCode::WriteMultipleRegisters, &mut cursor).unwrap() {
+                    Request::WriteMultipleRegisters(write) => write,
+                    _ => panic!("bad match"),
+                };
+
+            assert_eq!(registers.range, AddressRange::try_from(1, 2).unwrap());
+            assert_eq!(
+                registers.iterator.collect::<Vec<Indexed<u16>>>(),
+                vec![Indexed::new(1, 0xCAFE), Indexed::new(2, 0xBBDD)]
+            )
         }
     }
 }
