+++ rodbus/src/service/mod.rs
@@ -1,5 +1,5 @@
+mod serialization;
+
 pub(crate) mod function;
 pub(crate) mod impls;
-pub(crate) mod parse;
-mod serialization;
 pub(crate) mod traits;
