+++ rodbus/src/service/serialization/mod.rs
@@ -1,3 +1,2 @@
 mod client_request_parsers;
-mod client_response_parsers;
 mod serialize;
